Separate representation phase, flatten representation picking
