+++ ffc/ir/uflacs/elementtables.py
@@ -14,7 +14,7 @@ import numpy
 import ufl
 import ufl.utils.derivativetuples
 from ffc.fiatinterface import create_element
-from ffc.representation.representationutils import (create_quadrature_points_and_weights,
+from ffc.ir.representationutils import (create_quadrature_points_and_weights,
                                                     integral_type_to_entity_dim,
                                                     map_integral_points)
