--- ffc/tsfc/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (C) 2016 Jan Blechta
-#
-# This file is part of FFC (https://www.fenicsproject.org)
-#
-# SPDX-License-Identifier:    LGPL-3.0-or-later
-
-from ffc.tsfc.tsfcrepresentation import compute_integral_ir  # noqa: F401
-from ffc.tsfc.tsfcgenerator import generate_integral_code  # noqa: F401
